Added base for root removal

+++ Unity_GGJ2023/Assets/Scripts/Rotatable.cs
@@ -2,12 +2,12 @@ using UnityEngine;
 public class Rotatable : MonoBehaviour, IRotatable
 {
     [SerializeField] private Transform rotatingObject;
-    public void AddRotation(int degrees)
+    public void AddRotation(float degrees)
     {
         rotatingObject.transform.Rotate(Vector3.back, degrees, Space.Self);
     }
 }
 public interface IRotatable
 {
-    void AddRotation(int degrees);
+    void AddRotation(float degrees);
 }
